+++ src/scripts/bingo2/core.js
@@ -8,6 +8,8 @@
         noop = function () { },
         slice = Array.prototype.slice;
 
+    var fpName = '_bg_ifFn_',spName = '_bg_ifStr_';
+
     var _htmlDivTarget = null,
     _getHtmlDivTarget = function () {
         return _htmlDivTarget || (_htmlDivTarget = document.createElement('div'));
@@ -19,7 +21,7 @@
 
     var bingo = window.bingo = {
         //主版本号.子版本号.修正版本号.编译版本号(日期)
-        version: { major: 2, minor: 0, rev: 2, build: 160814, toString: function () { return [this.major, this.minor, this.rev, this.build].join('.'); } },
+        version: { major: 2, minor: 0, rev: 2, build: 160816, toString: function () { return [this.major, this.minor, this.rev, this.build].join('.'); } },
         bgNoObserve: true,//防止observe
         isDebug: false,
         prdtVersion: '',
@@ -60,14 +62,17 @@
             return (this.isNull(s) || s === stringEmpty);
         },
         isFunction: function (fun) {
-            return this.isType("Function", fun);
+            (fpName in Function.prototype) || (Function.prototype[fpName] = true);
+            return !this.isNull(fun) && fun[fpName] === true;
         },
         isNumeric: function (n) {
             //return this.isType("Number", n) && !isNaN(n) && isFinite(n);;
             return !isNaN(parseFloat(n)) && isFinite(n);
         },
         isString: function (obj) {
-            return this.isType("String", obj);
+            (spName in String.prototype) || (String.prototype[spName] = true);
+            return !this.isNull(obj) && obj[spName] === true;
+            //return this.isType("String", obj);
         },
         isObject: function (obj) {
             return !this.isNull(obj) && this.isType("Object", obj)